fix: resolve "TypeError: loader is not a function" with AP_EXECUTION_MODE=SANDBOXED
Move env-migrations to @activepieces/server-utils so the engine can
reuse the SANDBOXED → SANDBOX_PROCESS migration that previously only
ran in the API layer. Also move DatabaseType and RedisType enums to
server-utils and add a guard for unknown execution mode values.
