+++ packages/server/engine/package.json
@@ -12,6 +12,7 @@
     "webpack-ignore-dynamic-require": "1.0.0",
     "@activepieces/pieces-common": "workspace:*",
     "@activepieces/pieces-framework": "workspace:*",
+    "@activepieces/server-utils": "workspace:*",
     "@activepieces/shared": "workspace:*",
     "ai": "^6.0.0",
     "async-mutex": "0.4.0",
