+++ packages/server/utils/src/database-type.ts
@@ -1,0 +1,4 @@
+export enum DatabaseType {
+    POSTGRES = 'POSTGRES',
+    PGLITE = 'PGLITE',
+}
