+++ packages/server/utils/src/env-migrations.ts
@@ -1,24 +1,27 @@
 import { ExecutionMode } from '@activepieces/shared'
-import { DatabaseType } from '../../database/database-type'
-import { RedisType } from '../../database/redis/types'
-import { AppSystemProp } from './system-props'
+import { DatabaseType } from './database-type'
+import { RedisType } from './redis-type'
 
-const envPrefix = (prop: string): string => `AP_${prop}`
+const ENV_VAR_NAMES = {
+    EXECUTION_MODE: 'AP_EXECUTION_MODE',
+    REDIS_TYPE: 'AP_REDIS_TYPE',
+    DB_TYPE: 'AP_DB_TYPE',
+    QUEUE_MODE: 'AP_QUEUE_MODE',
+}
 
 export const environmentMigrations = {
     migrate(): Record<string, string | undefined> {
-
         return {
             ...process.env,
-            [envPrefix(AppSystemProp.EXECUTION_MODE)]: migrateExecutionMode(getRawValue(AppSystemProp.EXECUTION_MODE)),
-            [envPrefix(AppSystemProp.REDIS_TYPE)]: migrateRedisType(getRawValue(AppSystemProp.REDIS_TYPE)),
-            [envPrefix(AppSystemProp.DB_TYPE)]: migrateDbType(getRawValue(AppSystemProp.DB_TYPE)),
+            [ENV_VAR_NAMES.EXECUTION_MODE]: migrateExecutionMode(process.env[ENV_VAR_NAMES.EXECUTION_MODE]),
+            [ENV_VAR_NAMES.REDIS_TYPE]: migrateRedisType(process.env[ENV_VAR_NAMES.REDIS_TYPE]),
+            [ENV_VAR_NAMES.DB_TYPE]: migrateDbType(process.env[ENV_VAR_NAMES.DB_TYPE]),
         }
     },
 }
 
 function migrateRedisType(currentRedisType: string | undefined): string | undefined {
-    const queueMode = process.env['AP_QUEUE_MODE']
+    const queueMode = process.env[ENV_VAR_NAMES.QUEUE_MODE]
     if (queueMode === 'MEMORY') {
         return RedisType.MEMORY
     }
@@ -37,8 +40,4 @@ function migrateDbType(currentDbType: string | undefined): string | undefined {
         return DatabaseType.PGLITE
     }
     return currentDbType
-}
-
-function getRawValue(prop: string): string | undefined {
-    return process.env[envPrefix(prop)]
 }
