+++ packages/server/utils/src/index.ts
@@ -1,7 +1,10 @@
 export { cryptoUtils } from './crypto'
+export { DatabaseType } from './database-type'
+export { environmentMigrations } from './env-migrations'
 export { onCallService } from './on-call.service'
 export { apDayjs, apDayjsDuration } from './dayjs-helper'
 export { fileSystemUtils, INFINITE_LOCK_TIMEOUT } from './file-system-utils'
 export { memoryLock } from './memory-lock'
 export type { ApLock } from './memory-lock'
+export { RedisType } from './redis-type'
 export { systemUsage } from './system-usage'
