+++ packages/server/utils/src/redis-type.ts
@@ -1,0 +1,5 @@
+export enum RedisType {
+    SENTINEL = 'SENTINEL',
+    MEMORY = 'MEMORY',
+    STANDALONE = 'STANDALONE',
+}
